+++ src/Symfony/Bundle/FrameworkBundle/Tests/Controller/ControllerResolverTest.php
@@ -13,74 +13,15 @@ namespace Symfony\Bundle\FrameworkBundle\Tests\Controller;
 
 use Psr\Container\ContainerInterface as Psr11ContainerInterface;
 use Psr\Log\LoggerInterface;
-use Symfony\Bridge\PhpUnit\ExpectDeprecationTrait;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Bundle\FrameworkBundle\Controller\ControllerResolver;
 use Symfony\Component\DependencyInjection\Container;
-use Symfony\Component\DependencyInjection\ContainerAwareInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Tests\Controller\ContainerControllerResolverTest;
 
 class ControllerResolverTest extends ContainerControllerResolverTest
 {
-    use ExpectDeprecationTrait;
-
-    /**
-     * @group legacy
-     */
-    public function testGetControllerOnContainerAware()
-    {
-        $resolver = $this->createControllerResolver();
-        $request = Request::create('/');
-        $request->attributes->set('_controller', 'Symfony\Bundle\FrameworkBundle\Tests\Controller\ContainerAwareController::testAction');
-
-        $this->expectDeprecation('Since symfony/dependency-injection 6.4: Relying on "Symfony\Component\DependencyInjection\ContainerAwareInterface" to get the container in "Symfony\Bundle\FrameworkBundle\Tests\Controller\ContainerAwareController" is deprecated, register the controller as a service and use dependency injection instead.');
-        $controller = $resolver->getController($request);
-
-        $this->assertInstanceOf(ContainerAwareController::class, $controller[0]);
-        $this->assertInstanceOf(ContainerInterface::class, $controller[0]->getContainer());
-        $this->assertSame('testAction', $controller[1]);
-    }
-
-    /**
-     * @group legacy
-     */
-    public function testGetControllerOnContainerAwareInvokable()
-    {
-        $resolver = $this->createControllerResolver();
-        $request = Request::create('/');
-        $request->attributes->set('_controller', 'Symfony\Bundle\FrameworkBundle\Tests\Controller\ContainerAwareController');
-
-        $this->expectDeprecation('Since symfony/dependency-injection 6.4: Relying on "Symfony\Component\DependencyInjection\ContainerAwareInterface" to get the container in "Symfony\Bundle\FrameworkBundle\Tests\Controller\ContainerAwareController" is deprecated, register the controller as a service and use dependency injection instead.');
-        $controller = $resolver->getController($request);
-
-        $this->assertInstanceOf(ContainerAwareController::class, $controller);
-        $this->assertInstanceOf(ContainerInterface::class, $controller->getContainer());
-    }
-
-    /**
-     * @group legacy
-     */
-    public function testContainerAwareControllerGetsContainerWhenNotSet()
-    {
-        class_exists(AbstractControllerTest::class);
-
-        $controller = new ContainerAwareController();
-
-        $container = new Container();
-        $container->set(TestAbstractController::class, $controller);
-
-        $resolver = $this->createControllerResolver(null, $container);
-
-        $request = Request::create('/');
-        $request->attributes->set('_controller', TestAbstractController::class.'::testAction');
-
-        $this->expectDeprecation('Since symfony/dependency-injection 6.4: Relying on "Symfony\Component\DependencyInjection\ContainerAwareInterface" to get the container in "Symfony\Bundle\FrameworkBundle\Tests\Controller\ContainerAwareController" is deprecated, register the controller as a service and use dependency injection instead.');
-        $this->assertSame([$controller, 'testAction'], $resolver->getController($request));
-        $this->assertSame($container, $controller->getContainer());
-    }
-
     public function testAbstractControllerGetsContainerWhenNotSet()
     {
         $this->expectException(\LogicException::class);
@@ -180,29 +121,6 @@ class ControllerResolverTest extends ContainerControllerResolverTest
     protected function createMockContainer()
     {
         return $this->createMock(ContainerInterface::class);
-    }
-}
-
-class ContainerAwareController implements ContainerAwareInterface
-{
-    private ?ContainerInterface $container = null;
-
-    public function setContainer(?ContainerInterface $container): void
-    {
-        $this->container = $container;
-    }
-
-    public function getContainer(): ?ContainerInterface
-    {
-        return $this->container;
-    }
-
-    public function testAction()
-    {
-    }
-
-    public function __invoke()
-    {
     }
 }