+++ src/Symfony/Component/DependencyInjection/Compiler/AutowirePass.php
@@ -15,7 +15,6 @@ use Symfony\Component\Config\Resource\ClassExistenceResource;
 use Symfony\Component\DependencyInjection\Attribute\Autowire;
 use Symfony\Component\DependencyInjection\Attribute\AutowireCallable;
 use Symfony\Component\DependencyInjection\Attribute\AutowireDecorated;
-use Symfony\Component\DependencyInjection\Attribute\MapDecorated;
 use Symfony\Component\DependencyInjection\Attribute\Target;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\ContainerInterface;
@@ -98,7 +97,7 @@ class AutowirePass extends AbstractRecursivePass
             return $this->processValue($this->container->getParameterBag()->resolveValue($value->value));
         }
 
-        if ($value instanceof AutowireDecorated || $value instanceof MapDecorated) {
+        if ($value instanceof AutowireDecorated) {
             $definition = $this->container->getDefinition($this->currentId);
 
             return new Reference($definition->innerServiceId ?? $this->currentId.'.inner', $definition->decorationOnInvalid ?? ContainerInterface::NULL_ON_INVALID_REFERENCE);
@@ -354,12 +353,6 @@ class AutowirePass extends AbstractRecursivePass
                 }
 
                 foreach ($parameter->getAttributes(AutowireDecorated::class) as $attribute) {
-                    $arguments[$index] = $this->processValue($attribute->newInstance());
-
-                    continue 2;
-                }
-
-                foreach ($parameter->getAttributes(MapDecorated::class) as $attribute) {
                     $arguments[$index] = $this->processValue($attribute->newInstance());
 
                     continue 2;