+++ src/Symfony/Component/DependencyInjection/Compiler/ServiceLocatorTagPass.php
@@ -64,28 +64,7 @@ final class ServiceLocatorTagPass extends AbstractRecursivePass
             throw new InvalidArgumentException(sprintf('Invalid definition for service "%s": an array of references is expected as first argument when the "container.service_locator" tag is set.', $this->currentId));
         }
 
-        $i = 0;
-
-        foreach ($services as $k => $v) {
-            if ($v instanceof ServiceClosureArgument) {
-                continue;
-            }
-
-            if ($i === $k) {
-                if ($v instanceof Reference) {
-                    unset($services[$k]);
-                    $k = (string) $v;
-                }
-                ++$i;
-            } elseif (\is_int($k)) {
-                $i = null;
-            }
-
-            $services[$k] = new ServiceClosureArgument($v);
-        }
-        ksort($services);
-
-        $value->setArgument(0, $services);
+        $value->setArgument(0, self::map($services));
 
         $id = '.service_locator.'.ContainerBuilder::hash($value);
 
@@ -104,12 +83,8 @@ final class ServiceLocatorTagPass extends AbstractRecursivePass
 
     public static function register(ContainerBuilder $container, array $map, string $callerId = null): Reference
     {
-        foreach ($map as $k => $v) {
-            $map[$k] = new ServiceClosureArgument($v);
-        }
-
         $locator = (new Definition(ServiceLocator::class))
-            ->addArgument($map)
+            ->addArgument(self::map($map))
             ->addTag('container.service_locator');
 
         if (null !== $callerId && $container->hasDefinition($callerId)) {
@@ -133,5 +108,31 @@ final class ServiceLocatorTagPass extends AbstractRecursivePass
         }
 
         return new Reference($id);
+    }
+
+    public static function map(array $services): array
+    {
+        $i = 0;
+
+        foreach ($services as $k => $v) {
+            if ($v instanceof ServiceClosureArgument) {
+                continue;
+            }
+
+            if ($i === $k) {
+                if ($v instanceof Reference) {
+                    unset($services[$k]);
+                    $k = (string) $v;
+                }
+                ++$i;
+            } elseif (\is_int($k)) {
+                $i = null;
+            }
+
+            $services[$k] = new ServiceClosureArgument($v);
+        }
+        ksort($services);
+
+        return $services;
     }
 }