+++ src/Symfony/Component/DependencyInjection/Loader/Configurator/ContainerConfigurator.php
@@ -125,10 +125,6 @@ function service_locator(array $values): ServiceLocatorArgument
 {
     $values = AbstractConfigurator::processValue($values, true);
 
-    if (isset($values[0])) {
-        trigger_deprecation('symfony/dependency-injection', '6.3', 'Using integers as keys in a "service_locator()" argument is deprecated. The keys will default to the IDs of the original services in 7.0.');
-    }
-
     return new ServiceLocatorArgument($values);
 }