+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/php/services_subscriber.php
@@ -43,10 +43,9 @@ class ProjectServiceContainer extends Container
     public function getRemovedIds(): array
     {
         return [
-            '.service_locator.0H1ht0q' => true,
-            '.service_locator.0H1ht0q.foo_service' => true,
             '.service_locator.2hyyc9y' => true,
             '.service_locator.KGUGnmw' => true,
+            '.service_locator.KGUGnmw.foo_service' => true,
             'Symfony\\Component\\DependencyInjection\\Tests\\Fixtures\\CustomDefinition' => true,
         ];
     }