+++ src/Symfony/Bridge/Doctrine/CHANGELOG.md
@@ -9,6 +9,7 @@ CHANGELOG
  * Remove `RememberMeTokenProviderDoctrineSchemaSubscriber`, use `RememberMeTokenProviderDoctrineSchemaListener` instead
  * Remove `DbalLogger`, use a middleware instead
  * Remove `DoctrineDataCollector::addLogger()`, use a `DebugDataHolder` instead
+ * Remove `ContainerAwareLoader`, use dependency injection in your fixtures instead
  * `ContainerAwareEventManager::getListeners()` must be called with an event name
  * DoctrineBridge now requires `doctrine/event-manager:^2`
  * Add parameter `$isSameDatabase` to `DoctrineTokenProvider::configureSchema()`